[6] fix: general UI details
