+++ src/components/Naviguation.svelte
@@ -37,102 +37,6 @@
           href: '/checkbox',
           text: 'Checkbox',
         },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
-        {
-          href: '/button',
-          text: 'Button',
-        },
-        {
-          href: '/checkbox',
-          text: 'Checkbox',
-        },
       ],
     },
   ];
